+++ src/main/java/model/piece/AbsoluteValueDecreaser.java
@@ -1,4 +1,4 @@
-package model;
+package model.piece;
 
 public final class AbsoluteValueDecreaser {
